Fixed the default preset export
I could not use this project due to the way the default export was done in the main `preset.js`.

I believe the issue is that storybook expects ESM for presets.

This appears to fix it for me at least in windows, if someone else could try it out on mac/linux, that would be great to confirm it does not break anything else.